face-recognition/requirements.txt
google-cloud-storage
google-cloud-vision
face_recognition
opencv-python
numpy

